[MC] Permit BuildAttribute values to be 64-bit
The current generic implementation stores values as unsigned. This
is sufficient for all existing uses.

The AArch64 PAuthABI attributes are defined to be 64-bit values
and the PAuthABI profile for bare-metal will use values > 32-bits
to store flags.

TODO update tests.

+++ llvm/include/llvm/MC/MCELFStreamer.h
@@ -85,9 +85,9 @@ public:
       NumericAndTextAttributes
     } Type;
     unsigned Tag;
-    unsigned IntValue;
+    uint64_t IntValue;
     std::string StringValue;
-    AttributeItem(Types Ty, unsigned Tg, unsigned IV, std::string SV)
+    AttributeItem(Types Ty, unsigned Tg, uint64_t IV, std::string SV)
         : Type(Ty), Tag(Tg), IntValue(IV), StringValue(std::move(SV)) {}
   };
 
@@ -102,11 +102,11 @@ public:
 
   // Attributes that are added and managed entirely by target.
   SmallVector<AttributeItem, 64> Contents;
-  void setAttributeItem(unsigned Attribute, unsigned Value,
+  void setAttributeItem(unsigned Attribute, uint64_t Value,
                         bool OverwriteExisting);
   void setAttributeItem(unsigned Attribute, StringRef Value,
                         bool OverwriteExisting);
-  void setAttributeItems(unsigned Attribute, unsigned IntValue,
+  void setAttributeItems(unsigned Attribute, uint64_t IntValue,
                          StringRef StringValue, bool OverwriteExisting);
   void emitAttributesSection(StringRef Vendor, const Twine &Section,
                              unsigned Type, MCSection *&AttributeSection) {
